requirements.txt
Django==2.2.12
django-filter==2.3.0